packages/@grc/ai/pyproject.toml
[project]
name = "grc-ai"
version = "0.2.0"
description = "Multi-cloud AI provider abstraction for AI Interview Tool"
requires-python = ">=3.12"
license = { text = "MIT" }

dependencies = [
    "pydantic>=2.10.0",
    "httpx>=0.28.0",
    "tenacity>=9.0.0",
    # OpenAI / Azure OpenAI (GPT-4o, GPT-4 Turbo, o1, o3)
    "openai>=1.59.0",
    # Anthropic Claude (Claude 3.5 Sonnet, Claude 3 Opus, Claude 3.5 Haiku)
    "anthropic>=0.42.0",
    # Azure Speech Services
    "azure-cognitiveservices-speech>=1.42.0",
    # AWS Bedrock (Claude, Llama 3, Titan)
    "boto3>=1.35.0",
    # GCP Vertex AI (Gemini 2.0, PaLM 2)
    "google-cloud-aiplatform>=1.75.0",
    "google-cloud-speech>=2.28.0",
    "google-cloud-texttospeech>=2.22.0",
    "google-cloud-translate>=3.18.0",
    # Token counting
    "tiktoken>=0.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/grc_ai"]
